requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
dj-database-url==2.2.0
dj-email-url==1.0.6
Django==5.1.2
django-cache-url==3.4.5
djangorestframework==3.15.2
environs==11.0.0
marshmallow==3.23.0
packaging==24.1
psycopg==3.2.3
psycopg-binary==3.2.3
python-dotenv==1.0.1
redis==5.2.0
sqlparse==0.5.1
typing_extensions==4.12.2
